+++ src/api/app.ts
@@ -19,6 +19,7 @@ import { pollRoutes } from "./routes/polls";
 import { signupSheetRoutes } from "./routes/signup-sheets";
 import { CALPACA_VERSION } from "../version";
 import { oneOffOfferRoutes } from "./routes/one-off-offers";
+import { availabilityTroubleshooterRoutes } from "./routes/availability-troubleshooter";
 
 export const app = new Hono();
 
@@ -63,3 +64,4 @@ app.route("/", inviteeCalendarRoutes);
 app.route("/", pollRoutes);
 app.route("/", signupSheetRoutes);
 app.route("/", oneOffOfferRoutes);
+app.route("/", availabilityTroubleshooterRoutes);